source adjust spaces

+++ TestNetwork/TestNetwork/TcpClient.cpp
@@ -24,11 +24,8 @@ namespace Simplenet {
     void TcpClient::process()
     {
         int s = socket(PF_INET6, SOCK_STREAM, IPPROTO_TCP);
-        
         int r = -1;
-        
         struct hostent* hp = gethostbyname2(_ip.c_str(), AF_INET6);
-        
         // set inet6
         struct sockaddr_in6 sin;
         memset(&sin, 0, sizeof(sin));
@@ -36,8 +33,6 @@ namespace Simplenet {
         memcpy((char *)&sin.sin6_addr, hp->h_addr, hp->h_length);
         sin.sin6_family = hp->h_addrtype ; // or AF_INET6 (address family)
         sin.sin6_port = htons(_port);
-      //  sin.sin6_addr = in6addr_any;
-        
         r = connect(s, (struct sockaddr *)&sin, sizeof(sin));
         assert(0<= r);
         while(!_exit) {
@@ -45,7 +40,7 @@ namespace Simplenet {
             std::this_thread::sleep_for(std::chrono::seconds(1));
             char buf[300];
             memset(buf, 0, sizeof(buf));
-            r = recv(s, buf, 300, 0);
+            r = (int)recv(s, buf, 300, 0);
             if (r>0) {
             printf("Receive %d Bytes = %s\n", r,buf);
             }
